api/requirements.txt
fastapi==0.115.12
instaloader==4.14.1
moviepy==2.1.2
pydantic==2.11.1
yt_dlp==2025.3.31
uvicorn==0.34.0
python-dotenv